requirements.txt
pymupdf
python-dotenv
sentence-transformers
chromadb
openai
streamlit
